feat : 펫 정보 등록 기능구현
