+++ app/src/main/java/com/example/togetherpet/Login/LoginActivityViewModel.kt
@@ -1,8 +1,9 @@
-package com.example.togetherpet
+package com.example.togetherpet.Login
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.togetherpet.TokenRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
