+++ app/src/main/java/com/example/togetherpet/Login/LoginService.kt
@@ -1,4 +1,4 @@
-package com.example.togetherpet
+package com.example.togetherpet.Login
 
 import retrofit2.http.Body
 import retrofit2.http.POST
