--- app/src/main/java/com/example/togetherpet/MainApplication.kt
@@ -1,13 +1,0 @@
-package com.example.togetherpet
-
-import android.app.Application
-import com.kakao.sdk.common.KakaoSdk
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class MainApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        KakaoSdk.init(this, BuildConfig.KAKAO_NATIVE_APP_KEY)
-    }
-}
