+++ app/src/main/java/com/example/togetherpet/Registration/RegistrationImageFragment.kt
@@ -1,4 +1,4 @@
-package com.example.togetherpet
+package com.example.togetherpet.Registration
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,20 +6,23 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.navigation.fragment.findNavController
+import androidx.fragment.app.activityViewModels
 import com.example.togetherpet.databinding.FragmentInfoRegistrationImageBinding
-import com.example.togetherpet.databinding.FragmentInfoRegistrationStartBinding
+import dagger.hilt.android.AndroidEntryPoint
 
 
-class RegistrationStartFragment : Fragment() {
-    private var binding : FragmentInfoRegistrationStartBinding? = null
+@AndroidEntryPoint
+class RegistrationImageFragment : Fragment() {
+    private var binding : FragmentInfoRegistrationImageBinding? = null
+    private val sharedViewModel : RegistrationViewModel by activityViewModels()
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        binding = FragmentInfoRegistrationStartBinding.inflate(inflater)
+        binding = FragmentInfoRegistrationImageBinding.inflate(inflater)
         return binding?.root
     }
 
@@ -27,17 +30,17 @@ class RegistrationStartFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         binding?.apply {
-            startWhiteButton.setOnClickListener { goToNextScreen() }
-            startPinkButton.setOnClickListener { goToFindScreen() }
+            nextButton.setOnClickListener { goToNextScreen() }
         }
     }
 
     private fun goToNextScreen(){
-        findNavController().navigate(R.id.action_registrationStartFragment_to_registrationPetFragment)
+        Toast.makeText(activity, "next", Toast.LENGTH_SHORT).show()
+        sharedViewModel.sendPetInfoToServer()
     }
 
-    private fun goToFindScreen(){
-        Toast.makeText(activity, "find", Toast.LENGTH_SHORT).show()
+    override fun onDestroyView() {
+        super.onDestroyView()
+        binding = null
     }
-
 }
