--- app/src/main/java/com/example/togetherpet/RegistrationImageFragment.kt
@@ -1,40 +1,0 @@
-package com.example.togetherpet
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Toast
-import com.example.togetherpet.databinding.FragmentInfoRegistrationImageBinding
-
-
-class RegistrationImageFragment : Fragment() {
-    private var binding : FragmentInfoRegistrationImageBinding? = null
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        binding = FragmentInfoRegistrationImageBinding.inflate(inflater)
-        return binding?.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding?.apply {
-            nextButton.setOnClickListener { goToNextScreen() }
-        }
-    }
-
-    private fun goToNextScreen(){
-        Toast.makeText(activity, "next", Toast.LENGTH_SHORT).show()
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        binding = null
-    }
-}
