+++ app/src/main/java/com/example/togetherpet/TokenDataSource.kt
@@ -1,5 +1,7 @@
 package com.example.togetherpet
 
+import com.example.togetherpet.Login.LoginResponseDto
+
 interface TokenDataSource {
 
     // todo : response 수정 필요
