+++ app/src/main/res/navigation/reg_navigation_graph.xml
@@ -6,11 +6,11 @@
 
     <fragment
         android:id="@+id/registrationImageFragment"
-        android:name="com.example.togetherpet.RegistrationImageFragment"
+        android:name="com.example.togetherpet.Registration.RegistrationImageFragment"
         android:label="RegistrationImageFragment" />
     <fragment
         android:id="@+id/registrationPetFragment"
-        android:name="com.example.togetherpet.RegistrationPetFragment"
+        android:name="com.example.togetherpet.Registration.RegistrationPetFragment"
         android:label="RegistrationPetFragment" >
         <action
             android:id="@+id/action_registrationPetFragment_to_registrationResidenceFragment"
@@ -18,7 +18,7 @@
     </fragment>
     <fragment
         android:id="@+id/registrationResidenceFragment"
-        android:name="com.example.togetherpet.RegistrationResidenceFragment"
+        android:name="com.example.togetherpet.Registration.RegistrationResidenceFragment"
         android:label="RegistrationResidenceFragment" >
         <action
             android:id="@+id/action_registrationResidenceFragment_to_registrationImageFragment"
@@ -26,7 +26,7 @@
     </fragment>
     <fragment
         android:id="@+id/registrationStartFragment"
-        android:name="com.example.togetherpet.RegistrationStartFragment"
+        android:name="com.example.togetherpet.Registration.RegistrationStartFragment"
         android:label="RegistrationStartFragment" >
         <action
             android:id="@+id/action_registrationStartFragment_to_registrationPetFragment"
